added a new test to triangulate marking concept: a board with one number already marked
This test forced me to write a complex set of statements just to extract the tokens from my input string. It can be seen as the extended split operation it replaces.

I didn't introduced new concepts and I feel a made a step toward the abstraction I'm looking for the board and the number.

Yet, I wonder if the step was not too big or if I could find another way to "extract" tokens from the input string using only variables and creating objects instead.
But, it's now explicit from the regex what is a bingo Number. This use of regex doesn't seem like the good thing.

Let's see how it goes