triangulating winning concept : interesting case
I'm not sure where this implementation leads. It's a conflict where I want my implementation code to tell the truth about the behavior but it was not the simplest implementation to make the test pass :

from one hand I wanted to express "both number need to be marked for the board to win" while "number 0 (or number 1 for that matter)" is enough to make the test pass.

What am I missing here?