winning/loosing refacto: attempt to refactor what I've learned so far.
I see a kind of recursive function that checks if all numbers are marked, which is good but not enough, for me, to call it a day.

I'll try to refactor it in a more concise way, doing very little steps and see where it leads.

Also, I feel I need more behaviors to correctly abstract this winning/loosing concept because I only doing it on a single row and the board is a grid.